Framework: Move date module to packages maintained by Lerna
