+++ core-blocks/latest-posts/edit.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import { get, isUndefined, pickBy } from 'lodash';
-import moment from 'moment';
 import classnames from 'classnames';
 import { stringify } from 'querystringify';
 
@@ -20,6 +19,7 @@ import {
 	Toolbar,
 	withAPIData,
 } from '@wordpress/components';
+import { moment } from '@wordpress/date';
 import { __ } from '@wordpress/i18n';
 import { decodeEntities } from '@wordpress/utils';
 import {
