+++ lerna.json
@@ -1,0 +1,12 @@
+{
+	"lerna": "2.11.0",
+	"commands": {
+		"publish": {
+			"message": "chore(release): publish %s"
+		}
+	},
+	"packages": [
+		"packages/*"
+	],
+	"version": "2.8.0"
+}
