+++ package.json
@@ -1,8 +1,7 @@
 {
 	"name": "gutenberg",
-	"version": "2.8.0",
+	"private": true,
 	"description": "A new WordPress editor experience",
-	"main": "build/app.js",
 	"repository": "git+https://github.com/WordPress/gutenberg.git",
 	"author": "The WordPress Contributors",
 	"license": "GPL-2.0-or-later",
@@ -36,8 +35,8 @@
 		"js-beautify": "1.6.14",
 		"lodash": "4.17.5",
 		"memize": "1.0.5",
-		"moment": "2.21.0",
-		"moment-timezone": "0.5.13",
+		"moment": "2.22.1",
+		"moment-timezone": "0.5.16",
 		"mousetrap": "1.6.1",
 		"prop-types": "15.5.10",
 		"querystringify": "1.0.0",
@@ -81,6 +80,7 @@
 		"eslint-plugin-jsx-a11y": "6.0.2",
 		"eslint-plugin-react": "7.7.0",
 		"extract-text-webpack-plugin": "4.0.0-beta.0",
+		"lerna": "2.11.0",
 		"node-sass": "4.7.2",
 		"pegjs": "0.10.0",
 		"pegjs-loader": "0.5.4",
@@ -134,7 +134,11 @@
 		"fixtures:generate": "npm run fixtures:server-registered && cross-env GENERATE_MISSING_FIXTURES=y npm run test-unit",
 		"fixtures:regenerate": "npm run fixtures:clean && npm run fixtures:generate",
 		"package-plugin": "./bin/build-plugin-zip.sh",
+		"postinstall": "lerna bootstrap --hoist",
 		"pot-to-php": "./bin/pot-to-php.js",
+		"publish:check": "lerna updated",
+		"publish:dev": "lerna publish --npm-tag next",
+		"publish:prod": "lerna publish",
 		"test-unit": "wp-scripts test-unit-js --config test/unit/jest.config.json",
 		"test-unit-php": "docker-compose run --rm wordpress_phpunit phpunit",
 		"test-unit-php-multisite": "docker-compose run -e WP_MULTISITE=1 --rm wordpress_phpunit phpunit",
