+++ packages/date/README.md
@@ -1,0 +1,11 @@
+# @wordpress/date
+
+Date module for WordPress.
+
+## Installation
+
+Install the module
+
+```bash
+npm install @wordpress/date --save
+```
