+++ packages/date/package.json
@@ -1,0 +1,26 @@
+{
+	"name": "@wordpress/date",
+	"version": "0.0.1",
+	"description": "Date module for WordPress",
+	"author": "WordPress",
+	"license": "GPL-2.0-or-later",
+	"keywords": [
+		"wordpress",
+		"date"
+	],
+	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/date/README.md",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/WordPress/gutenberg.git"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/gutenberg/issues"
+	},
+	"dependencies": {
+		"moment": "2.22.1",
+		"moment-timezone": "0.5.16"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
