+++ test/unit/jest.config.json
@@ -1,10 +1,12 @@
 {
 	"rootDir": "../../",
 	"collectCoverageFrom": [
-		"(blocks|components|date|editor|element|data|utils|edit-post|viewport|plugins|core-data|core-blocks)/**/*.js"
+		"(blocks|components|editor|element|data|utils|edit-post|viewport|plugins|core-data|core-blocks)/**/*.js",
+		"packages/**/*.js"
 	],
 	"moduleNameMapper": {
-		"@wordpress\\/(blocks|components|date|editor|element|data|utils|edit-post|viewport|plugins|core-data|core-blocks)": "$1"
+		"@wordpress\\/(blocks|components|editor|element|data|utils|edit-post|viewport|plugins|core-data|core-blocks)": "$1",
+		"@wordpress\\/(date)": "packages/$1"
 	},
 	"preset": "@wordpress/jest-preset-default",
 	"setupFiles": [
